WebSockets: split channel.js into channel.js and message.js
